Changed traversal of dependency graph results so that the `ArtifactSet` instances included in the result are visited in consumer first order, rather than the nodes, so that the artifact sets can be collected directly into a list in the correct order rather than collecting and retaining an additional map from node -> artifact set to apply the ordering.

+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/SelectedArtifactResults.java
@@ -30,5 +30,5 @@ public interface SelectedArtifactResults {
     /**
      * Returns the set of artifacts with the given id.
      */
-    ResolvedArtifactSet getArtifactsWithId(long id);
+    ResolvedArtifactSet getArtifactsWithId(int id);
 }
