+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/oldresult/DefaultResolvedConfigurationBuilder.java
@@ -52,7 +52,7 @@ public class DefaultResolvedConfigurationBuilder implements ResolvedConfiguratio
     }
 
     @Override
-    public void addChild(DependencyGraphNode parent, DependencyGraphNode child, long artifactsId) {
+    public void addChild(DependencyGraphNode parent, DependencyGraphNode child, int artifactsId) {
         builder.parentChildMapping(parent.getNodeId(), child.getNodeId(), artifactsId);
     }
