+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/oldresult/ResolvedConfigurationBuilder.java
@@ -26,7 +26,7 @@ public interface ResolvedConfigurationBuilder {
 
     void addUnresolvedDependency(UnresolvedDependency unresolvedDependency);
 
-    void addChild(DependencyGraphNode parent, DependencyGraphNode child, long artifactsId);
+    void addChild(DependencyGraphNode parent, DependencyGraphNode child, int artifactsId);
 
     void newResolvedDependency(DependencyGraphNode node);
